agregados iconos de bootstrap y agregado barra de busqueda en ALumnosm mejorada funcion de eliminar alumnos, mensaje de eror si se intenta eliminar un alumno con expediente

+++ ColegioMVC/Controllers/AlumnosController.cs
@@ -139,12 +139,25 @@ namespace ColegioMVC.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var alumno = await _context.Alumno.FindAsync(id);
+
+            // Verificar si tiene expedientes
+            bool tieneExpedientes = _context.Expediente
+                .Any(e => e.AlumnoId == id);
+
+            if (tieneExpedientes)
+            {
+                TempData["Error"] =
+                    "No se puede eliminar el alumno porque tiene expedientes registrados.";
+
+                return RedirectToAction(nameof(Index));
+            }
+
             if (alumno != null)
             {
                 _context.Alumno.Remove(alumno);
+                await _context.SaveChangesAsync();
             }
 
-            await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
